requirements.txt
Flask>=2.0
requests>=2.0
bcrypt>=4.0
pytz>=2020.1
pytest>=7.0
python-dotenv>=1.0
